Add the script to convert python files to .ipynb
This commit adds a py2nb.py script which allows to convert python
scripts written with special comments to .ipynb notebooks.

Currently it supports Markdown cells (which are created from comments
starting with `#: `) and python code cells (any non-special lines)

By default consecutive lines of the same type are inserted into the same
cell, but a new cell can be forcefully created by inserting a special
separator comment `#=`.

To execute the code cells and store their outputs in the notebook, the
`-e` flag must be provided, otherwise the `execution_count` field of
each cell will be `0` and the `outputs` list will be empty. If an
exception occurs during the execution of a cell, the traceback is
appended to the `stderr` output of the cell and the execution of any
subsequent cells is aborted.